Get info and settings only once, assist user w/ conf, various fixes
I apologize for a quite messy patch. Some things were worked on in
parallel here for various reasons.

This patch will verify that a client configuration exists before
proceeding, and will offer to help create one if a machine with a
corresponding name is found.

This patch will also verify that there is an actual machine
corresponding to the specified client.